+++ KFXUtilities/Classes/KFXGeoLocationHelper.h
@@ -54,7 +54,14 @@ NS_ASSUME_NONNULL_BEGIN
 -(NSArray<CLLocation*>*)sortLocations:(NSArray<CLLocation*>*)locations
                  byDistanceToLocation:(CLLocation*)theLocation
                             ascending:(BOOL)ascending;
--(NSArray<id> *)sortObjects:(NSArray<id> *)objects withLocationKey:(NSString *)key byDistanceToLocation:(CLLocation *)theLocation ascending:(BOOL)ascending;
-
+-(NSArray<id> *)sortObjects:(NSArray<id> *)objects
+            withLocationKey:(NSString *)key
+       byDistanceToLocation:(CLLocation *)theLocation
+                  ascending:(BOOL)ascending;
+-(NSArray<id> *)sortObjects:(NSArray<id> *)objects
+            withLocationKey:(NSString *)key
+       byDistanceToLocation:(CLLocation *)theLocation
+                  ascending:(BOOL)ascending
+                  distances:(NSArray<NSNumber*>*__autoreleasing  _Nonnull*_Nonnull)distances;
 @end
 NS_ASSUME_NONNULL_END